app/requirements.txt
fastapi
uvicorn
requests
pydub
python-multipart
sentence-transformers
faiss-cpu